refactor(editor): move plugin toolbar from core to editor-ui
