+++ src/serlo-editor/editor-ui/plugin-toolbar/dropdown/default-controls.tsx
