+++ src/serlo-editor/editor-ui/plugin-toolbar/index.ts
