+++ src/serlo-editor/editor-ui/plugin-toolbar/plugin-toolbar-dropdown-menu.tsx
