+++ src/serlo-editor/editor-ui/plugin-toolbar/plugin-toolbar.tsx
