+++ src/serlo-editor/editor-ui/plugin-toolbar/text-controls/color-text-icon.tsx
