+++ src/serlo-editor/editor-ui/plugin-toolbar/text-controls/hooks/use-formatting-options.tsx
