+++ src/serlo-editor/editor-ui/plugin-toolbar/text-controls/plugin-toolbar-text-controls.tsx
