+++ src/serlo-editor/editor-ui/plugin-toolbar/text-controls/types.ts
