+++ src/serlo-editor/editor-ui/plugin-toolbar/text-controls/utils/color.ts
