+++ src/serlo-editor/plugins/text/components/math-element.tsx
@@ -10,7 +10,7 @@ import {
 import { MathFormula } from './math-formula'
 import type { MathElement as MathElementType, Paragraph } from '../types'
 import { PreferenceContext } from '@/serlo-editor/core'
-import { isElementWithinList } from '@/serlo-editor/core/plugin-toolbar/text-controls/utils/list'
+import { isElementWithinList } from '@/serlo-editor/editor-ui/plugin-toolbar/text-controls/utils/list'
 import { MathEditor } from '@/serlo-editor/math'
 
 export interface MathElementProps {
