+++ src/serlo-editor/plugins/text/types/config.ts
@@ -1,8 +1,8 @@
 import { Dispatch, SetStateAction } from 'react'
 import { Descendant, Editor, Range } from 'slate'
 
-import { useFormattingOptions } from '@/serlo-editor/core/plugin-toolbar/text-controls/hooks/use-formatting-options'
-import { TextEditorFormattingOption } from '@/serlo-editor/core/plugin-toolbar/text-controls/types'
+import { useFormattingOptions } from '@/serlo-editor/editor-ui/plugin-toolbar/text-controls/hooks/use-formatting-options'
+import { TextEditorFormattingOption } from '@/serlo-editor/editor-ui/plugin-toolbar/text-controls/types'
 import { SerializedScalarStateType } from '@/serlo-editor/plugin'
 
 export type TextEditorState = SerializedScalarStateType<
